pdf_data_extraction_agent/requirements.txt
google-adk>=1.15.0
google-genai>=1.39.0
pandas
python-dotenv
numpy
